improve how links are opened when clicked fix #231
When the user opens a link in the webkit view the request is catched and
send to the default browser then the click is ignored in webkit.
This solves the error users get when they click links, when Marker runs
inside flatpak (without network access), because this way Marker doesn't make
a network connection, see #231.